fix: component 이동
